+++ backend/coffeeing/src/main/java/com/ssafy/coffeeing/modules/product/service/CapsuleReviewService.java
@@ -79,7 +79,7 @@ public class CapsuleReviewService {
             throw new BusinessException(AuthErrorInfo.UNAUTHORIZED);
         }
 
-        review.update(reviewRequest.content(), reviewRequest.score());
+        review.update(reviewRequest.content(), (double) reviewRequest.score());
     }
 
     public void deleteReview(Long id) {
